Anpassungen

+++ whisper_flask.py
@@ -12,9 +12,12 @@ app = Flask(__name__)
 #tokenizer = WhisperTokenizer.from_pretrained("facebook/whisper-large")
 model = whisper.load_model("medium")
 
-@app.route('/', methods=['GET', 'POST'])
+@app.route('/')
 def index():
-    if request.method == 'POST':
+    return render_template('index.html')
+
+@app.route('/transcribe', methods=['POST'])
+def transcribe():
         # Überprüfen, ob eine Datei hochgeladen wurde
         if 'file' not in request.files:
             return render_template('index.html', error='Keine Datei ausgewählt')
@@ -24,28 +27,34 @@ def index():
             return render_template('index.html', error='Keine Datei ausgewählt')
 
         # Speichern Sie die hochgeladene Datei vorübergehend
-        audio_file_path = tempfile.mktemp(suffix=".wav")
+        audio_file_path = tempfile.mktemp(suffix=".mp3")
         file.save(audio_file_path)
 
         # Transkribieren Sie die Audiodatei
         options = {"language": "de"}
-        transcription_text = model.transcribe(audio_file_path, **options)
+        response = model.transcribe(audio_file_path, **options)
 
-        # Text in eine temporäre Datei schreiben
-        output_file_path = tempfile.mktemp(suffix=".txt")
-        with open(output_file_path, 'w') as output_file:
-            output_file.write(transcription_text)
+        # Extrahieren Sie die Transkription aus der Antwort
+        transcription_text = response.get("transcription")
 
-        # Datei zum Download bereitstellen
-        response = send_file(output_file_path, as_attachment=True, download_name='transcription.txt')
+        if transcription_text:
+            # Text in eine temporäre Datei schreiben
+            output_file_path = tempfile.mktemp(suffix=".txt")
+            with open(output_file_path, 'w') as output_file:
+                output_file.write(transcription_text)
 
-        # Hochgeladene Audiodatei und temporäre Textdatei löschen
-        os.remove(audio_file_path)
-        os.remove(output_file_path)
+            # Datei zum Download bereitstellen
+            response = send_file(output_file_path, as_attachment=True, download_name='transcription.txt')
 
-        return response
+            # Hochgeladene Audiodatei und temporäre Textdatei löschen
+            os.remove(audio_file_path)
+            os.remove(output_file_path)
 
-    return render_template('index.html')
+            return response
+        else:
+            # Wenn keine Transkription vorhanden ist, geben Sie einen Fehler zurück
+            return render_template('index.html', error='Transkription fehlgeschlagen', warning='Transkription erfolgreich, aber es gab Warnungen oder Hinweise während des Prozesses')
+
 
 
 if __name__ == '__main__':
